# docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample swagger for Fiber",
        "title": "TES SWAGGER ULBI",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/ghaidafasya24",
            "email": "714220031@std.ulbi.ac.id"
        },
        "version": "1.0"
    },
    "host": "bp-tubes-c48fa88ca6a5.herokuapp.com",
    "basePath": "/",
    "paths": {
        "/insert": {
            "post": {
                "description": "Input data menu.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu"
                ],
                "summary": "Insert data menu.",
                "parameters": [
                    {
                        "description": "Payload Body [RAW]",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.ReqMenu"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Menu"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/restoran": {
            "get": {
                "description": "Mengambil semua data menu.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu"
                ],
                "summary": "Get All Data Menu.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Menu"
                        }
                    }
                }
            }
        },
        "/restoran/{id}": {
            "get": {
                "description": "Ambil per ID data menu.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu"
                ],
                "summary": "Get By ID Data Menu.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Masukan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Menu"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.BahanBaku": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string",
                    "example": "123456789"
                },
                "bahan_baku": {
                    "type": "string",
                    "example": "Nasi, Bumbu, Ayam"
                },
                "jumlah": {
                    "type": "string",
                    "example": "3"
                }
            }
        },
        "controller.Kategori": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string",
                    "example": "123456789"
                },
                "kategori": {
                    "type": "string",
                    "example": "Makanan"
                }
            }
        },
        "controller.Menu": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string",
                    "example": "123456789"
                },
                "bahan_baku": {
                    "$ref": "#/definitions/controller.BahanBaku"
                },
                "deskripsi": {
                    "type": "string",
                    "example": "Nasi goreng dengan campuran sayuran dan ayam"
                },
                "harga": {
                    "type": "number",
                    "example": 15000
                },
                "kategori": {
                    "$ref": "#/definitions/controller.Kategori"
                },
                "nama": {
                    "type": "string",
                    "example": "Nasi Goreng"
                }
            }
        },
        "controller.ReqBahanBaku": {
            "type": "object",
            "properties": {
                "bahan_baku": {
                    "type": "string",
                    "example": "Nasi, Bumbu, Ayam"
                },
                "jumlah": {
                    "type": "string",
                    "example": "3"
                }
            }
        },
        "controller.ReqKategori": {
            "type": "object",
            "properties": {
                "kategori": {
                    "type": "string",
                    "example": "Makanan"
                }
            }
        },
        "controller.ReqMenu": {
            "type": "object",
            "properties": {
                "bahan_baku": {
                    "$ref": "#/definitions/controller.ReqBahanBaku"
                },
                "deskripsi": {
                    "type": "string",
                    "example": "Nasi goreng dengan campuran sayuran dan ayam"
                },
                "harga": {
                    "type": "number",
                    "example": 15000
                },
                "kategori": {
                    "$ref": "#/definitions/controller.ReqKategori"
                },
                "nama": {
                    "type": "string",
                    "example": "Nasi Goreng"
                }
            }
        }
    }
}